@@ -1,26 +1,31 @@
 package com.thunder.shadowmanager;
 
-import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
-import android.os.Handler;
-import android.os.Looper;
 import android.os.RemoteException;
 import android.text.TextUtils;
+import android.util.Log;
+import android.util.Pair;
+
+import com.tencent.shadow.core.manager.installplugin.InstallPluginException;
 import com.tencent.shadow.core.manager.installplugin.InstalledPlugin;
 import com.tencent.shadow.core.manager.installplugin.InstalledType;
 import com.tencent.shadow.core.manager.installplugin.PluginConfig;
 import com.tencent.shadow.dynamic.host.EnterCallback;
 import com.tencent.shadow.dynamic.host.FailedException;
 import com.tencent.shadow.dynamic.manager.PluginManagerThatUseDynamicLoader;
+
 import org.json.JSONException;
+
 import java.io.File;
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
@@ -37,9 +42,8 @@ import java.util.concurrent.TimeoutException;
 public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
 
     private ExecutorService executorService = Executors.newSingleThreadExecutor();
-    private ExecutorService executorService2 = Executors.newSingleThreadExecutor();
     private ExecutorService mFixedPool = Executors.newFixedThreadPool(4);
-    private String mProcessServiceName ="";
+    private String mProcessServiceName = "";
 
     public MyPluginManager(Context context) {
         super(context);
@@ -58,28 +62,6 @@ public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
     /**
      * TODO 下面内容需要自己实现
      *
-     * @return demo插件so的abi
-     */
-    @Override
-    public String getAbi() {
-        return "";
-    }
-
-    /**
-     * TODO 下面内容需要自己实现
-     *
-     * @return 宿主中注册的PluginProcessService实现的类名
-     */
-    protected String getPluginProcessServiceName() {
-        return mProcessServiceName;
-    }
-    protected void setPluginProcessServiceName(String pluginProcessServiceName){
-        this.mProcessServiceName =pluginProcessServiceName;
-    }
-
-    /**
-     * TODO 下面内容需要自己实现
-     *
      * @param context  context
      * @param fromId   标识本次请求的来源位置，用于区分入口
      * @param bundle   参数列表
@@ -91,27 +73,23 @@ public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
 
         final String partKey = bundle.getString("part_key");
         mProcessServiceName = bundle.getString("process_service_name");
-       String  appVersion = bundle.getString(" app_version");
-        final String className = "com.k.b.MainActivity";
+        String appVersion = bundle.getString(" app_version");
+        final String className = "com.tencent.shadow.sample.plugin.runtime.PluginDefaultProxyActivity";
         final String pluginZipPath = bundle.getString("p_p");
-//        if (className == null) {
-//            throw new NullPointerException("className == null");
-//        }
-
         if (fromId == 1011) { // 打开 Activity 示例
             final Bundle extras = bundle.getBundle("extra_to_plugin_bundle");
             if (callback != null) {
                 // 开始加载插件了，实现加载布局
                 callback.onShowLoadingView(null);
             }
-
             executorService.execute(new Runnable() {
                 @Override
                 public void run() {
                     try {
                         InstalledPlugin installedPlugin = installPlugin(context, pluginZipPath, null, true);//这个调用是阻塞的
+                        loadPlugin(installedPlugin.UUID, partKey);
                         Intent pluginIntent = new Intent();
-                         pluginIntent.putExtra("app_version",appVersion);
+                        pluginIntent.putExtra("app_version", appVersion);
                         pluginIntent.setClassName(
                                 context.getPackageName(),
                                 className
@@ -120,19 +98,17 @@ public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
                             pluginIntent.replaceExtras(extras);
                         }
 
-                        startPluginActivity(context, installedPlugin, partKey, pluginIntent);
+                        Intent intent = mPluginLoader.convertActivityIntent(pluginIntent);
+                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                        mPluginLoader.startActivityInPluginProcess(intent);
                     } catch (Exception e) {
                         throw new RuntimeException(e);
                     }
                     if (callback != null) {
-                        Handler uiHandler = new Handler(Looper.getMainLooper());
-                        uiHandler.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                // 到这里插件就启动完成了
-                                callback.onCloseLoadingView();
-                                callback.onEnterComplete();
-                            }
+                        mUiHandler.post(() -> {
+                            // 到这里插件就启动完成了
+                            callback.onCloseLoadingView();
+                            callback.onEnterComplete();
                         });
                     }
                 }
@@ -141,12 +117,11 @@ public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
         }
     }
 
-
     public InstalledPlugin installPlugin(Context context, String zip, String hash, boolean odex) throws IOException, JSONException, InterruptedException, ExecutionException {
         final PluginConfig pluginConfig = installPluginFromZip(new File(zip), hash);
         final String uuid = pluginConfig.UUID;
-
         List<Future> futures = new LinkedList<>();
+        List<Future<Pair<String, String>>> extractSoFutures = new LinkedList<>();
         if (pluginConfig.runTime != null && pluginConfig.pluginLoader != null) {
             Future odexRuntime = mFixedPool.submit(new Callable() {
                 @Override
@@ -170,15 +145,16 @@ public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
         for (Map.Entry<String, PluginConfig.PluginFileInfo> plugin : pluginConfig.plugins.entrySet()) {
             final String partKey = plugin.getKey();
             final File apkFile = plugin.getValue().file;
-            Future extractSo = mFixedPool.submit((Callable) () -> {
-                extractSo(uuid, partKey, apkFile);
-                return null;
-            });
+            Future<Pair<String, String>> extractSo = mFixedPool.submit(() -> extractSo(uuid, partKey, apkFile));
             futures.add(extractSo);
+            extractSoFutures.add(extractSo);
             if (odex) {
-                Future odexPlugin = mFixedPool.submit((Callable) () -> {
-                    oDexPlugin(uuid, partKey, apkFile);
-                    return null;
+                Future odexPlugin = mFixedPool.submit(new Callable() {
+                    @Override
+                    public Object call() throws Exception {
+                        oDexPlugin(uuid, partKey, apkFile);
+                        return null;
+                    }
                 });
                 futures.add(odexPlugin);
             }
@@ -187,41 +163,19 @@ public class MyPluginManager extends PluginManagerThatUseDynamicLoader {
         for (Future future : futures) {
             future.get();
         }
-        onInstallCompleted(pluginConfig);
-
-        SharedPreferences p_ver = context.getSharedPreferences("p_ver", Context.MODE_PRIVATE);
-        String puuid = p_ver.getString("puuid", "");
-        if (!TextUtils.isEmpty(puuid)) {
-            if (!TextUtils.equals(puuid, uuid)) {
-                deleteInstalledPlugin(puuid);
-                p_ver.edit().putString("puuid", uuid).apply();
-            }
-        } else {
-            p_ver.edit().putString("puuid", uuid).apply();
+        Map<String, String> soDirMap = new HashMap<>();
+        for (Future<Pair<String, String>> future : extractSoFutures) {
+            Pair<String, String> pair = future.get();
+            soDirMap.put(pair.first, pair.second);
         }
+        onInstallCompleted(pluginConfig, soDirMap);
 
         return getInstalledPlugins(1).get(0);
     }
 
-
-    public void startPluginActivity(Context context, InstalledPlugin installedPlugin, String partKey, Intent pluginIntent) throws RemoteException, TimeoutException, FailedException {
-        Intent intent = convertActivityIntent(installedPlugin, partKey, pluginIntent);
-        if (!(context instanceof Activity)) {
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-        }
-
-        context.startActivity(intent);
-
-    }
-
-    public Intent convertActivityIntent(InstalledPlugin installedPlugin, String partKey, Intent pluginIntent) throws RemoteException, TimeoutException, FailedException {
-        loadPlugin(installedPlugin.UUID, partKey);
-        return mPluginLoader.convertActivityIntent(pluginIntent);
-    }
-
     private void loadPluginLoaderAndRuntime(String uuid) throws RemoteException, TimeoutException, FailedException {
         if (mPpsController == null) {
-            bindPluginProcessService(getPluginProcessServiceName());
+            bindPluginProcessService(mProcessServiceName);
             waitServiceConnected(10, TimeUnit.SECONDS);
         }
         loadRunTime(uuid);
